% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseTracks.R
\name{browseTracks}
\alias{browseTracks}
\title{browse tracks}
\usage{
browseTracks(
  tracks,
  locus,
  genome,
  formats = "auto",
  additionalFiles,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{tracks}{filenames for signals to import.}

\item{locus}{genomic location to show.}

\item{genome}{genome assembly.}

\item{formats}{format of the files.}

\item{additionalFiles}{additional files for the tracks. 
Must use same name as tracks.}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{...}{parameters could be used by 
igv.js(https://github.com/igvteam/igv.js/wiki/Tracks-2.0).}
}
\value{
An object of class htmlwidget that will intelligently print itself 
into HTML in a variety of contexts including the R console, 
within R Markdown documents, and within Shiny output bindings.
}
\description{
browse tracks by a web browser.
}
\examples{
url <- "https://www.encodeproject.org/files/ENCFF356YES/@download/ENCFF356YES.bigWig"
browseTracks(url, genome="hg19", locus="chr11:122929275-122930122")
}
